+++ GeoNorge/AT.Common.GeoNorge.Publish/Implementation/AdresserRequestAdapter.cs
@@ -1,0 +1,12 @@
+using Microsoft.Kiota.Abstractions.Authentication;
+using Microsoft.Kiota.Bundle;
+
+namespace Arbeidstilsynet.Common.GeoNorge.Implementation;
+
+internal class AdresserRequestAdapter(IHttpClientFactory httpClientFactory)
+    : DefaultRequestAdapter(
+        new AnonymousAuthenticationProvider(),
+        httpClient: httpClientFactory.CreateClient(
+            DependencyInjection.DependencyInjectionExtensions.AdresserHttpClientName
+        )
+    ) { }