+++ GeoNorge/AT.Common.GeoNorge.Test/Integration/AddressClientIntegrationTests.cs
@@ -9,7 +9,6 @@ namespace Arbeidstilsynet.Common.GeoNorge.Adapters.Test.Integration;
 public class AddressClientIntegrationTests : TestBed<GeoNorgeTestFixture>
 {
     private readonly IAddressSearch _sut;
-    private readonly VerifySettings _verifySettings = new VerifySettings();
 
     public AddressClientIntegrationTests(
         ITestOutputHelper testOutputHelper,
@@ -18,38 +17,16 @@ public class AddressClientIntegrationTests : TestBed<GeoNorgeTestFixture>
         : base(testOutputHelper, fixture)
     {
         _sut = fixture.GetService<IAddressSearch>(testOutputHelper)!;
-
-        _verifySettings.UseDirectory("TestData/Snapshots");
     }
 
     [Fact]
-    public async Task SearchAddresses_ValidRequest_DeserializesResult()
+    public async Task SearchAddresses_ValidRequest_MapsToDomainResult()
     {
         // Act
-        var result = await _sut.SearchAddresses(
-            new TextSearchQuery() { SearchTerm = "Storgata 1, Oslo" }
-        );
+        var result = await _sut.SearchAddresses(new TextSearchQuery { SearchTerm = "Storgata 1" });
 
         // Assert
         result.ShouldNotBeNull();
-        await Verify(result, _verifySettings);
-    }
-
-    [Fact]
-    public async Task SearchAddressesByPoint_ValidRequest_DeserializesResult()
-    {
-        // Act
-        var result = await _sut.SearchAddressesByPoint(
-            new PointSearchQuery()
-            {
-                Latitude = 4.2,
-                Longitude = 4.2,
-                RadiusInMeters = 42,
-            }
-        );
-
-        // Assert
-        result.ShouldNotBeNull();
-        await Verify(result, _verifySettings);
+        result.Adresser.ShouldNotBeNull();
     }
 }