+++ GeoNorge/AT.Common.GeoNorge.Test/Unit/AddressSearchExtensionsTests.cs
@@ -1,6 +1,6 @@
+using Arbeidstilsynet.Common.GeoNorge.Adresser.Models;
 using Arbeidstilsynet.Common.GeoNorge.Extensions;
 using Arbeidstilsynet.Common.GeoNorge.Model.Request;
-using Arbeidstilsynet.Common.GeoNorge.Model.Response;
 using Arbeidstilsynet.Common.GeoNorge.Ports;
 using NSubstitute;
 using Shouldly;
@@ -42,22 +42,22 @@ public class AddressSearchExtensionsTests
             RadiusInMeters = 1000,
         };
 
-        var returnedElements = new List<Address>()
+        var returnedList = new OutputGeoPointList
         {
-            new() { Adressenavn = "Testveien 1" },
-            new() { Adressenavn = "Testveien 2" },
+            Adresser =
+            [
+                new OutputGeoPoint { Adressenavn = "Testveien 1" },
+                new OutputGeoPoint { Adressenavn = "Testveien 2" },
+            ],
         };
 
-        var returnedPagination = new PaginationResult<Address>() { Elements = returnedElements };
-
-        _addressSearch.SearchAddressesByPoint(default!).ReturnsForAnyArgs(returnedPagination);
+        _addressSearch.SearchAddressesByPoint(default!).ReturnsForAnyArgs(returnedList);
 
         // Act
         var result = await _addressSearch.GetClosestAddress(query);
 
         // Assert
-
-        result.ShouldBe(returnedPagination.Elements.FirstOrDefault());
+        result.ShouldBe(returnedList.Adresser.FirstOrDefault());
     }
 
     [Fact]
@@ -95,27 +95,27 @@ public class AddressSearchExtensionsTests
             Kommunenummer = "2510",
         };
 
-        var returnedElements = new List<Address>()
+        var returnedList = new OutputAdresseList
         {
-            new()
-            {
-                Location = new Location() { Latitude = 60.0, Longitude = 10.0 },
-            },
-            new()
-            {
-                Location = new Location() { Latitude = 61.0, Longitude = 11.0 },
-            },
+            Adresser =
+            [
+                new OutputAdresse
+                {
+                    Representasjonspunkt = new GeomPoint { Lat = 60.0f, Lon = 10.0f },
+                },
+                new OutputAdresse
+                {
+                    Representasjonspunkt = new GeomPoint { Lat = 61.0f, Lon = 11.0f },
+                },
+            ],
         };
 
-        var returnedPagination = new PaginationResult<Address>() { Elements = returnedElements };
-
-        _addressSearch.SearchAddresses(default!, default!).ReturnsForAnyArgs(returnedPagination);
+        _addressSearch.SearchAddresses(default!, default!).ReturnsForAnyArgs(returnedList);
 
         // Act
         var result = await _addressSearch.QuickSearchLocation(query);
 
         // Assert
-
-        result.ShouldBe(returnedPagination.Elements.FirstOrDefault()?.Location);
+        result.ShouldBe(returnedList.Adresser.FirstOrDefault()?.Representasjonspunkt);
     }
 }